Packaging: -Added dependencies with autoconf archive macro imported in unisim_lib/m4
PowerPC G4:
-Fixed lot of bugs to make it run again Linux


Former-commit-id: 8da22ce0b7567c399f3971b853ec33870c9b5ce5

+++ unisim_lib/unisim/component/cxx/memory/flash/am29/am29lv800b_config.hh
@@ -47,7 +47,7 @@ namespace am29 {
 class AM29LV800BConfig
 {
 public:
-	typedef uint32_t ADDRESS;
+	typedef uint64_t ADDRESS;
 	typedef enum
 	{
 		ST_I0 = 0,
